+++ .php-cs-fixer.dist.php
@@ -18,6 +18,6 @@ return (new Config())
     ->setFinder(
         (new Finder())
             ->in(__DIR__)
-            ->exclude('example')
+            ->exclude(['example', 'tests/fixtures'])
     )
 ;